backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pymupdf>=1.24.0
openai>=1.40.0
pydantic>=2.8.0
python-dotenv>=1.0.1
